pmoves/services/media-video/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
ultralytics==8.2.103
Pillow==10.4.0
boto3==1.34.162
requests==2.32.3
torch==2.4.0
